1:ユーザー＿時間割

+++ database/migrations/2024_12_06_172911_create_delivery_times_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('delivery_times', function (Blueprint $table) {
             $table->id();
-            $table->integer('curriculums_id');
+            $table->unsignedBigInteger('curriculums_id');
             $table->dateTime('delivery_from');
             $table->dateTime('delivery_to');
             $table->timestamps();
